requirements.txt
pyserial == 3.5
adafruit-ampy == 1.1.0
